+++ CHANGELOG.md
@@ -15,3 +15,17 @@ This project attempts to adhere to [Semantic Versioning](https://semver.org/spec
   It implements encoding and decoding of data from and to Markdown formatted representation.
 - Package `keepachangelog`.
   It implements types and functions to assist in maintaining a Changelog based on the [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) format.
+- Command `get`, which shows changes for a specific version(s).
+- Command `version`, which prints the app version.
+- Command `promote`, which promotes unreleased draft to be the next release version.
+- Command `prepare`, which prepares the changelog for the next release cycle.
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security